added two env vars DHCE_HOST and DHCE_PORT to support running tests against remote DHCE servers, also refactored the unit tests to be more friendly to use for CI
